+++ bitmart/api_broker.py
@@ -4,7 +4,7 @@ from bitmart.lib.cloud_consts import *
 
 class APIBroker(CloudClient):
 
-    def __init__(self, api_key: str = "", secret_key: str = "", memo: str = "", url: str = API_URL, timeout: tuple = TIMEOUT, logger=None):
+    def __init__(self, api_key: str = "", secret_key: str = "", memo: str = "", url: str = API_URL, timeout: tuple = TIMEOUT, headers=None, logger=None):
         """
         Create api key from https://www.bitmart.com/api-config/en-US
         :param api_key: your access key
@@ -13,7 +13,7 @@ class APIBroker(CloudClient):
         :param url: https://api-cloud.bitmart.com
         :param timeout: (2, 10)
         """
-        CloudClient.__init__(self, api_key, secret_key, memo, url, timeout, logger)
+        CloudClient.__init__(self, api_key, secret_key, memo, url, timeout, headers, logger)
 
     # GET https://api-cloud.bitmart.com/spot/v1/broker/rebate
     def broker_rebates(self):