+++ bitmart/api_spot.py
@@ -7,7 +7,7 @@ from bitmart.lib.cloud_consts import *
 class APISpot(CloudClient):
 
     def __init__(self, api_key: str = "", secret_key: str = "", memo: str = "", url: str = API_URL,
-                 timeout: tuple = TIMEOUT, logger=None):
+                 timeout: tuple = TIMEOUT, headers=None, logger=None):
         """
         Create api key from https://www.bitmart.com/api-config/en-US
         :param api_key: your access key
@@ -16,7 +16,7 @@ class APISpot(CloudClient):
         :param url: https://api-cloud.bitmart.com
         :param timeout: (2, 10)
         """
-        CloudClient.__init__(self, api_key, secret_key, memo, url, timeout, logger)
+        CloudClient.__init__(self, api_key, secret_key, memo, url, timeout, headers, logger)
 
     # basic API
     def get_currencies(self):