+++ examples/spot/market/get_curreny_list.py
@@ -8,18 +8,10 @@ from bitmart.lib.cloud_utils import config_logging
 config_logging(logging, logging.DEBUG)
 logger = logging.getLogger(__name__)
 
-spotAPI = APISpot(api_key="Your_Api_Key",
-                  secret_key="Your_Secret_Key",
-                  memo="Your_Memo",
-                  logger=logger)
+spotAPI = APISpot(logger=logger)
 
 try:
-    response = spotAPI.post_submit_order(
-        symbol='BTC_USDT',
-        side='buy',
-        type='limit',
-        size='0.01',
-        price='8800')[0]
+    response = spotAPI.get_currencies()[0]
     logger.info(response)
 except APIException as error:
     logger.error(